# backend/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Job Hunting SaaS API
  version: "0.1.0"
  description: 就活管理SaaSのREST API
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/entries/{entryId}/selection-flow:
    parameters:
      - $ref: "#/components/parameters/EntryId"
    get:
      operationId: getSelectionFlow
      summary: エントリーに紐づく可変選考フローを取得する
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SelectionFlowResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertSelectionFlow
      summary: エントリーに紐づく可変選考フローを作成または置換する
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertSelectionFlowRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SelectionFlowResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/entries/{entryId}/selection-flow/current-stage:
    parameters:
      - $ref: "#/components/parameters/EntryId"
    patch:
      operationId: updateSelectionFlowCurrentStage
      summary: 可変選考フローの現在ステージを更新する
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSelectionFlowCurrentStageRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SelectionFlowResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/inbox/clips:
    post:
      operationId: createInboxClip
      summary: Chrome 拡張等から保存されたページクリップを作成する
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInboxClipRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxClipResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listInboxClips
      summary: 自分のページクリップ一覧を返す
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  clips:
                    type: array
                    items:
                      $ref: "#/components/schemas/InboxClipResponse"
                required:
                  - clips
  /api/v1/inbox/clips/{clipId}:
    parameters:
      - $ref: "#/components/parameters/ClipId"
    delete:
      operationId: deleteInboxClip
      summary: ページクリップを削除する
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
...
    EntryId:
      name: entryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ClipId:
      name: clipId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    SelectionStageInput:
      type: object
      properties:
        stageKind:
          type: string
          enum: [application, document, test, interview, group, offer, other]
        stageLabel:
          type: string
        evidenceText:
          type: string
      required:
        - stageKind
        - stageLabel
    UpsertSelectionFlowRequest:
      type: object
      properties:
        source:
          type: string
          enum: [template, manual, ai_inbox, ai_paste]
        currentStagePosition:
          type: integer
          minimum: 1
        confidence:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
        inboxClipId:
          type: string
          format: uuid
          nullable: true
        stages:
          type: array
          minItems: 1
          maxItems: 20
          items:
            $ref: "#/components/schemas/SelectionStageInput"
      required:
        - source
        - stages
    UpdateSelectionFlowCurrentStageRequest:
      type: object
      properties:
        position:
          type: integer
          minimum: 1
      required:
        - position
    SelectionStageResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        position:
          type: integer
        stageKind:
          type: string
        stageLabel:
          type: string
        evidenceText:
          type: string
      required:
        - id
        - position
        - stageKind
        - stageLabel
        - evidenceText
    SelectionFlowResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entryId:
          type: string
          format: uuid
        source:
          type: string
        currentStagePosition:
          type: integer
        confidence:
          type: integer
          nullable: true
        inboxClipId:
          type: string
          format: uuid
          nullable: true
        stages:
          type: array
          items:
            $ref: "#/components/schemas/SelectionStageResponse"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - entryId
        - source
        - currentStagePosition
        - stages
        - createdAt
        - updatedAt
...
    CreateInboxClipRequest:
      type: object
      properties:
        url:
          type: string
          maxLength: 2048
        title:
          type: string
          maxLength: 512
        source:
          type: string
          maxLength: 128
        guess:
          type: string
          maxLength: 256
        contentText:
          type: string
          maxLength: 20000
      required:
        - url
        - title
        - source
    InboxClipResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
        title:
          type: string
        source:
          type: string
        guess:
          type: string
        contentText:
          type: string
        capturedAt:
          type: string
          format: date-time
      required:
        - id
        - url
        - title
        - source
        - guess
        - contentText
        - capturedAt
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message